Health, Ferocity, Bug Fix, Focus Update.
Health and Energy stats are in the game. They don't do anything yet.
Ferocity is in as well, and even slowly drains, but also does nothing.
Fixed a bug where looking straight down caused the player to stop moving; fixed it by change the up and down degrees from 90 to 89 each.
Updated the gun to focus in when clutched.